# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rest API for catalog",
        "version": "1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "description": "User's credentials",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "leopold.kuhic@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get a JWT via given credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwczovL3BvcnRhbHRlc3QyLnJ1L2FwaS9hdXRoL2xvZ2luIiwiaWF0IjoxNzM2ODUwNTM5LCJleHAiOjE3MzY4NTQxMzksIm5iZiI6MTczNjg1MDUzOSwianRpIjoiRWZ5U1liTGhJMWlMbVNYSCIsInN1YiI6IjEiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.9766eqWuUCqI-993e6nCPohNrQfUFk6v1DRDtqoFqLA"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/buildings": {
            "get": {
                "tags": [
                    "Building"
                ],
                "summary": "Get list of buildings",
                "description": "Returns list of buildings",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ApiResourceCollection"
                                            }
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Building"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/auth/buildings/map/point/{latitude}/{longitude}/{radius}": {
            "get": {
                "tags": [
                    "Building"
                ],
                "summary": "Get list of Buildings by map point and radius",
                "description": "Returns list of buildings by map point and radius",
                "operationId": "getBuildingsByMapPoint",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "path",
                        "description": "latitude",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "47.897709"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "path",
                        "description": "longitude",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "40.082366"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "path",
                        "description": "radius (km)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ApiResourceCollection"
                                            }
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Building"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/auth/buildings/rectangle/{latitude1}/{longitude1}/{latitude2}/{longitude2}": {
            "get": {
                "tags": [
                    "Building"
                ],
                "summary": "Get list of buildings by rectangle",
                "description": "Returns list of buildings by rectangle",
                "operationId": "getBuildingByRectangle",
                "parameters": [
                    {
                        "name": "latitude1",
                        "in": "path",
                        "description": "latitude1",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "47.897709"
                        }
                    },
                    {
                        "name": "longitude1",
                        "in": "path",
                        "description": "longitude1",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "40.082366"
                        }
                    },
                    {
                        "name": "latitude2",
                        "in": "path",
                        "description": "latitude2",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "48.897709"
                        }
                    },
                    {
                        "name": "longitude2",
                        "in": "path",
                        "description": "longitude2",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "41.082366"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ApiResourceCollection"
                                            }
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Building"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResourceCollection": {
                "properties": {
                    "total": {
                        "type": "integer",
                        "example": 10
                    },
                    "perPage": {
                        "type": "integer",
                        "example": 10
                    },
                    "currentPage": {
                        "type": "integer",
                        "example": 1
                    },
                    "totalPages": {
                        "type": "integer",
                        "example": 1
                    },
                    "nextPage": {
                        "type": "integer",
                        "example": null
                    },
                    "prevPage": {
                        "type": "integer",
                        "example": null
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ApiResourceCollection"
                }
            },
            "Building": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 47.897709
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": "40.082366"
                    },
                    "distance": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "67408 Frami Plains Apt. 051\n         * North Linnie, DC 98000-2914"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Building"
                }
            }
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}